250312 - ok so notebook is parsing all variables from bash script as expected now. The bug is in the output directory. Making empty output_directory (with proper subdirectories, also empty). Output is being saved inside the input_directory in a _results folder. Goal for today: redirect output.
--> ok actually, last I did was to remove the _results directory. That put me inside the input dir where the image files are. Will reuse the _results dir to maintain the file hierarchy. will up one more and call output.

--> input: ok so, for running as. jupyter notebook we can call the image_subdirecotry input_directory directly because there is only one image.
When running as array we should copy the path to the folder_name (local_download_directory) on script, so we can create output_directories for each, since we are now working with multiple images in parallel.

Still need to push these comments to Github. I have to move the notebook and script to the pl becasue my home dir is full, probably because of all the log files. cannot delete so need to contact the alpine people for halp.